Introduce Tide Trident
The Tide Trident is a throwable weapon that is created from Sea Serpent Fangs, Sea Serpent Scales, a Diamond, and a Dragon Bone. This trident is fairly strong, has support for the Return enchantment in Spartan Weaponry, and pierces enemies.

Additionally, this commit also allows the Podium to display any items, and fixes the dispenser behavior with regards to the different arrows.